+++ src/components/map-layer/o-map-layer.factory.ts
@@ -5,7 +5,7 @@ import * as L from 'leaflet';
 
 export class OMapLayerFactory {
 
-  public createMapLayer(layerConf: LayerConfiguration, mapService: MapService): L.ILayer {
+  public createMapLayer(layerConf: LayerConfiguration, mapService: MapService): L.Layer {
     let layer;
     switch (layerConf.type) {
       case 'marker':
@@ -38,9 +38,9 @@ export class OMapLayerFactory {
 
   /**
    * Creates marker and adds it to map.
-   * @param  {LayerConfiguration} layerConf
+   * @param  layerConf
    *          Layer configuration parameters.
-   * @param  {MapService} mapService
+   * @param  mapService
    *          Reference to map service responsible of interact with leaflet map object.
    * @return L.Marker layer.
    */
@@ -62,9 +62,9 @@ export class OMapLayerFactory {
 
   /**
    * Creates GeoJSON layer and adds it to map.
-   * @param  {LayerConfiguration} layerConf
+   * @param  layerConf
    *          Layer configuration parameters.
-   * @param  {MapService} mapService
+   * @param  mapService
    *          Reference to map service responsible of interact with leaflet map object.
    * @returns L.GeoJSON layer.
    */
@@ -95,11 +95,11 @@ export class OMapLayerFactory {
 
   /**
    * Creates WMS layer and adds it to map.
-   * @param  {LayerConfiguration} layerConf
+   * @param   layerConf
    *          Layer configuration parameters.
-   * @param  {MapService} mapService
+   * @param   mapService
    *          Reference to map service responsible of interact with leaflet map object.
-   * @returns L.TileLayer.WMS layer
+   * @returns  layer
    */
   public createWMS(layerConf: LayerConfiguration, mapService: MapService): L.TileLayer.WMS {
     var id = null;