+++ src/services.ts
@@ -1,17 +1,17 @@
 import { Injector } from '@angular/core';
-import { MdIconRegistry } from '@angular/material';
-import { TranslateMapService } from './TranslateMapService';
+import { MatIconRegistry } from '@angular/material';
+import { TranslateMapService } from './services/TranslateMapService';
 
-export * from './MapService';
-export * from './GeocodingService';
-export * from './TranslateMapService';
+export * from './services/MapService';
+export * from './services/GeocodingService';
+export * from './services/TranslateMapService';
 
 export function getTranslateMapServiceProvider(injector) {
   return new TranslateMapService(injector);
 }
 
 export const OMAP_PROVIDERS: any = [
-  MdIconRegistry,
+  MatIconRegistry,
   {
     provide: TranslateMapService,
     useFactory: getTranslateMapServiceProvider,